Move all editor logic out of lapce-core and into floem-editor-core
