+++ floem-editor-core/src/buffer/diff.rs
