+++ floem-editor-core/src/buffer/rope_text.rs
@@ -1,14 +1,8 @@
 use std::{borrow::Cow, ops::Range};
 
 use lapce_xi_rope::{interval::IntervalBounds, rope::ChunkIter, Cursor, Rope};
-use lsp_types::Position;
 
-use crate::{
-    encoding::{offset_utf16_to_utf8, offset_utf8_to_utf16},
-    mode::Mode,
-    paragraph::ParagraphCursor,
-    word::WordCursor,
-};
+use crate::{mode::Mode, paragraph::ParagraphCursor, word::WordCursor};
 
 pub trait RopeText {
     fn text(&self) -> &Rope;
@@ -47,39 +41,6 @@ pub trait RopeText {
             .unwrap_or(offset);
 
         self.text().line_of_offset(offset)
-    }
-
-    /// Converts a UTF8 offset to a UTF16 LSP position
-    /// Returns None if it is not a valid UTF16 offset
-    fn offset_to_position(&self, offset: usize) -> Position {
-        let (line, col) = self.offset_to_line_col(offset);
-        let line_offset = self.offset_of_line(line);
-
-        let utf16_col =
-            offset_utf8_to_utf16(self.char_indices_iter(line_offset..), col);
-
-        Position {
-            line: line as u32,
-            character: utf16_col as u32,
-        }
-    }
-
-    fn offset_of_position(&self, pos: &Position) -> usize {
-        let (line, column) = self.position_to_line_col(pos);
-
-        self.offset_of_line_col(line, column)
-    }
-
-    fn position_to_line_col(&self, pos: &Position) -> (usize, usize) {
-        let line = pos.line as usize;
-        let line_offset = self.offset_of_line(line);
-
-        let column = offset_utf16_to_utf8(
-            self.char_indices_iter(line_offset..),
-            pos.character as usize,
-        );
-
-        (line, column)
     }
 
     fn offset_to_line_col(&self, offset: usize) -> (usize, usize) {
