+++ floem-editor-core/src/char_buffer.rs
