+++ floem-editor-core/src/chars.rs
