+++ floem-editor-core/src/lib.rs
@@ -1,0 +1,15 @@
+pub mod buffer;
+pub mod char_buffer;
+pub mod chars;
+pub mod command;
+pub mod cursor;
+pub mod editor;
+pub mod indent;
+pub mod mode;
+pub mod movement;
+pub mod paragraph;
+pub mod register;
+pub mod selection;
+pub mod soft_tab;
+pub mod util;
+pub mod word;
