+++ floem-editor-core/src/mode.rs
