+++ floem-editor-core/src/selection.rs
