+++ floem-editor-core/src/word.rs
