+++ lapce-app/Cargo.toml
@@ -51,6 +51,7 @@ sled = "0.34.7"
 bytemuck = "1.14.0"
 tokio = { version = "1.21", features = ["full"] }
 futures = "0.3.26"
+floem-editor-core = { path = "../floem-editor-core" }
 floem = { git = "https://github.com/lapce/floem", rev = "6ae536237716d606e5546edb9f72867f2bc23946" }
 # floem = { path = "../../workspaces/floem" }
 floem-editor = { path = "../floem-editor" }
