+++ lapce-core/Cargo.toml
@@ -9,7 +9,6 @@ anyhow.workspace = true
 directories.workspace = true
 itertools.workspace = true
 once_cell.workspace = true
-serde.workspace = true
 strum.workspace = true
 strum_macros.workspace = true
 thiserror.workspace = true
@@ -22,8 +21,9 @@ lapce-xi-rope.workspace = true
 
 lapce-rpc.workspace = true
 
+floem-editor-core = { path = "../floem-editor-core" }
+
 libloading = "0.7.3"
-bitflags = "1.3.2"
 slotmap = "1.0"
 arc-swap = "1.6.0"
 tree-sitter = "0.20.10"
