+++ lapce-core/src/lib.rs
@@ -1,24 +1,13 @@
 #![allow(clippy::manual_clamp)]
 
-pub mod buffer;
-pub mod char_buffer;
-pub mod chars;
-pub mod command;
-pub mod cursor;
 pub mod directory;
-pub mod editor;
 pub mod encoding;
-pub mod indent;
 pub mod language;
 pub mod lens;
 pub mod meta;
-pub mod mode;
-pub mod movement;
-pub mod paragraph;
-pub mod register;
-pub mod selection;
-pub mod soft_tab;
+pub mod rope_text_pos;
 pub mod style;
 pub mod syntax;
-pub mod util;
-pub mod word;
+// This is primarily being re-exported to avoid changing every single usage
+// in lapce-app. We should probably remove this at some point.
+pub use floem_editor_core::*;
