+++ lapce-core/src/syntax/highlight.rs
@@ -22,8 +22,8 @@ use tree_sitter::{
     Language, Point, Query, QueryCaptures, QueryCursor, QueryMatch, Tree,
 };
 
-use super::PARSER;
-use crate::{language::LapceLanguage, style::SCOPES, util::RopeProvider};
+use super::{util::RopeProvider, PARSER};
+use crate::{language::LapceLanguage, style::SCOPES};
 
 thread_local! {
     static HIGHLIGHT_CONFIGS: RefCell<HashMap<LapceLanguage, Result<Arc<HighlightConfiguration>, HighlightIssue>>> = Default::default();
