+++ lapce-core/src/syntax/mod.rs
@@ -14,6 +14,7 @@ use std::{
     sync::{atomic::AtomicUsize, Arc},
 };
 
+use floem_editor_core::util::{matching_bracket_general, matching_pair_direction};
 use itertools::Itertools;
 use lapce_rpc::style::{LineStyle, Style};
 use lapce_xi_rope::{
@@ -31,18 +32,19 @@ use self::{
         HighlightConfiguration, HighlightEvent, HighlightIter, HighlightIterLayer,
         IncludedChildren, LocalScope,
     },
+    util::RopeProvider,
 };
 use crate::{
     buffer::rope_text::RopeText,
     language::{self, LapceLanguage},
     lens::{Lens, LensBuilder},
     style::SCOPES,
-    util::{matching_bracket_general, matching_pair_direction, RopeProvider},
 };
 
 use crate::buffer::Buffer;
 pub mod edit;
 pub mod highlight;
+pub mod util;
 
 // Uses significant portions Helix's implementation, and on tree-sitter's highlighter implementation
